Struct tbd::log::LogEntry

pub struct LogEntry<T: Hashable> {
    pub hash: Hash,
    pub dttm: Tm,
    pub entry: T,
    pub parent: Box<ParentEntry<T>>,
}

An single entry plus timestamp, its hash and its parent.

Its hash value is a sumary of the entry value, the timestamp and its parent hash.

Fields

hash

dttm

entry

parent

Methods

impl<T: Hashable> LogEntry<T>

fn entry_hash(&self) -> Hash