Module tbd::log

Logging library.

Structs

Log
Main log structure which points to the latest entry.

LogEntry
An single entry plus timestamp, its hash and its parent.

LogIterator

Enums

Hash
Stores one of the supported hash values.

ParentEntry
Refers to a precending entry or to the root.

Traits

Hashable
Defines if a type is able to generate a crypto hash which represents its content.

LogTrait
Functions a Log must provide to count as a log.

Readable

Writable

Functions

bin_slice_to_hex

save_head_to_fs

save_to_fs

tm_from_i64

tm_to_bytes
Convert a Tm time structure to a byte array.